requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
python-dotenv>=1.1.0
requests>=2.32.4
pytest>=8.4.0
pytest-django>=4.11.1
paypalrestsdk>=1.13.1 